location search moved into components directory
