+++ app/components/forecastfield.js
@@ -5,11 +5,11 @@ class ForecastField extends React.Component {
     constructor() {
       super();
     }
-
   render() {
     return (
       <section>
         Hey There! {this.props.data[0].date}
+        
       </section>
     )
   }
